+++ tests/integration_tests/adapters/betfair/test_betfair_common.py
@@ -12,10 +12,18 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
+
+from decimal import Decimal
+
+import pytest
+
 from nautilus_trader.adapters.betfair.common import BETFAIR_TICK_SCHEME
 from nautilus_trader.adapters.betfair.common import MAX_BET_PRICE
 from nautilus_trader.adapters.betfair.common import MIN_BET_PRICE
 from nautilus_trader.adapters.betfair.orderbook import betfair_float_to_price
+from nautilus_trader.model.objects import Price
+from nautilus_trader.model.objects import Quantity
+from tests.integration_tests.adapters.betfair.test_kit import betting_instrument
 
 
 class TestBetfairCommon:
@@ -29,3 +37,48 @@ class TestBetfairCommon:
     def test_betfair_ticks(self):
         assert self.tick_scheme.min_price == betfair_float_to_price(1.01)
         assert self.tick_scheme.max_price == betfair_float_to_price(1000)
+
+
+class TestBettingInstrument:
+    def setup(self):
+        self.instrument = betting_instrument()
+
+    def test_notional_value(self):
+        notional = self.instrument.notional_value(
+            quantity=Quantity.from_int(100),
+            price=Price.from_str("0.5"),
+            use_quote_for_inverse=False,
+        ).as_decimal()
+        # We are long 100 at 0.5 probability, aka 2.0 in odds terms
+        assert notional == Decimal("200.0")
+
+    @pytest.mark.parametrize(
+        ("value", "n", "expected"),
+        [
+            (101, 0, "110"),
+        ],
+    )
+    def test_next_ask_price(self, value, n, expected):
+        result = self.instrument.next_ask_price(value, num_ticks=n)
+        expected = Price.from_str(expected)
+        assert result == expected
+
+    @pytest.mark.parametrize(
+        ("value", "n", "expected"),
+        [
+            (1.999, 0, "1.99"),
+        ],
+    )
+    def test_next_bid_price(self, value, n, expected):
+        result = self.instrument.next_bid_price(value, num_ticks=n)
+        expected = Price.from_str(expected)
+        assert result == expected
+
+    def test_min_max_price(self):
+        assert self.instrument.min_price == Price.from_str("1.01")
+        assert self.instrument.max_price == Price.from_str("1000")
+
+    def test_to_dict(self):
+        instrument = betting_instrument()
+        data = instrument.to_dict(instrument)
+        assert data["venue_name"] == "BETFAIR"
